+++ src/gettsim/germany/grundsicherung/im_alter/einkommen.py
@@ -19,7 +19,6 @@ if TYPE_CHECKING:
 
 
 @policy_function(
-    start_date="2011-01-01",
     end_date="2017-12-31",
     leaf_name="einkommen_m",
 )
@@ -38,6 +37,8 @@ def einkommen_m_bis_2017(
     familie__anzahl_personen_sn: int,
     sozialversicherung__beiträge_versicherter_m: float,
     elterngeld__anrechenbarer_betrag_m: float,
+    unterhalt__tatsächlich_erhaltener_betrag_m: float,
+    unterhaltsvorschuss__betrag_m: float,
 ) -> float:
     """Income considered for Grundsicherung im Alter before 2018.
 
@@ -54,6 +55,8 @@ def einkommen_m_bis_2017(
         + einkommensteuer__einkünfte__aus_vermietung_und_verpachtung__betrag_m
         + kapitaleinkommen_brutto_m
         + elterngeld__anrechenbarer_betrag_m
+        + unterhalt__tatsächlich_erhaltener_betrag_m
+        + unterhaltsvorschuss__betrag_m
     )
 
     out = (
@@ -80,6 +83,8 @@ def einkommen_m_ab_2018(
     familie__anzahl_personen_sn: int,
     sozialversicherung__beiträge_versicherter_m: float,
     elterngeld__anrechenbarer_betrag_m: float,
+    unterhalt__tatsächlich_erhaltener_betrag_m: float,
+    unterhaltsvorschuss__betrag_m: float,
 ) -> float:
     """Income considered for Grundsicherung im Alter from 2018.
 
@@ -96,6 +101,8 @@ def einkommen_m_ab_2018(
         + einkommensteuer__einkünfte__aus_vermietung_und_verpachtung__betrag_m
         + kapitaleinkommen_brutto_m
         + elterngeld__anrechenbarer_betrag_m
+        + unterhalt__tatsächlich_erhaltener_betrag_m
+        + unterhaltsvorschuss__betrag_m
     )
 
     out = (